requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
attrs==19.1.0
backcall==0.1.0
cchardet==2.1.4
cffi==1.12.3
chardet==3.0.4
decorator==4.4.0
idna==2.8
idna-ssl==1.1.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
multidict==4.5.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pony==0.7.10
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
pycares==3.0.0
pycparser==2.19
Pygments==2.3.1
PyYAML==5.1
six==1.12.0
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
wcwidth==0.1.7
yarl==1.3.0
